+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Change Log
 
 ## Unreleased
+* Add NpmRunner class and npm-update-diff exectuable (#1)